requirements.txt
Flask==3.1.1
python-telegram-bot==20.8
tensorflow==2.18.0
scikit-learn==1.6.1